+++ FacturapiClient.cs
@@ -6,18 +6,18 @@ using System.Text;
 
 namespace Facturapi
 {
-    public sealed class FacturapiClient : IFacturapiClient
+    public sealed class FacturapiClient : IDisposable
     {
-        public ICustomerWrapper Customer { get; private set; }
-        public IProductWrapper Product { get; private set; }
-        public IInvoiceWrapper Invoice { get; private set; }
-        public IOrganizationWrapper Organization { get; private set; }
-        public IReceiptWrapper Receipt { get; private set; }
-        public IRetentionWrapper Retention { get; private set; }
-        public ICatalogWrapper Catalog { get; private set; }
-        public ICartaporteCatalogWrapper CartaporteCatalog { get; private set; }
-        public IToolWrapper Tool { get; private set; }
-        public IWebhookWrapper Webhook { get; private set; }
+        public CustomerWrapper Customer { get; private set; }
+        public ProductWrapper Product { get; private set; }
+        public InvoiceWrapper Invoice { get; private set; }
+        public OrganizationWrapper Organization { get; private set; }
+        public ReceiptWrapper Receipt { get; private set; }
+        public RetentionWrapper Retention { get; private set; }
+        public CatalogWrapper Catalog { get; private set; }
+        public CartaporteCatalogWrapper CartaporteCatalog { get; private set; }
+        public ToolWrapper Tool { get; private set; }
+        public WebhookWrapper Webhook { get; private set; }
         private readonly HttpClient httpClient;
         private readonly bool ownsHttpClient;
         private bool disposed;